requirements.txt
llvmlite==0.44.0
numba==0.61.0
numpy==1.24.0
openai
python-dotenv
langchain-community
langchain>=0.2.17
langchain-openai>=0.0.5
langchain-core>=0.2.43,<0.3.0
langchain-text-splitters
streamlit
